warnings removed & g. comp. casting rules review
Ignore-this: 6af1e3e2dd5f73d02e04f5e7d4a27963

darcs-hash:20100414154235-48308-45428a03c5ab9d1b5e28b8492f1b69561abbf9fe.gz

+++ kernel/tr/structures/system_tables.cpp
@@ -223,7 +223,7 @@ get_errors(xptr node,const char* /* title */)
     xptr parent = insert_element_i(XNULL,XNULL,node,"errors",xs_untyped,NULL_XMLNS);
     xptr left   = XNULL;
 
-    for (int i=0; i < user_error_code_entries_size/sizeof(user_error_code_entry); i++)
+    for (uint64_t i=0; i < user_error_code_entries_size/sizeof(user_error_code_entry); i++)
     {
         if (left==XNULL)
         {
